Combine chapter 7 and chapter 8.
